Move the model-table-delegate code to its place
