+++ activity_browser/layouts/tabs/activity.py
@@ -1,17 +1,14 @@
 # -*- coding: utf-8 -*-
-from copy import copy
-from dataclasses import dataclass
-from math import nan
-from typing import Any, Optional, Union
+from typing import Optional
 from peewee import DoesNotExist
-from PySide2 import QtCore, QtWidgets, QtGui
+from PySide2 import QtCore, QtWidgets
 from PySide2.QtCore import Slot
 
-from activity_browser import ab_settings, application, project_settings, signals
+from activity_browser import ab_settings, project_settings, signals
 from activity_browser.bwutils import commontasks as bc
-from activity_browser.logger import log
 from activity_browser.mod import bw2data as bd
-from activity_browser.ui.style import style_item
+from activity_browser.ui.tables.models.properties import PropertyModel
+from activity_browser.ui.tables.properties import PropertyTable
 
 from ...ui.icons import qicons
 from ...ui.style import style_activity_tab
@@ -21,7 +18,6 @@ from ...ui.tables import (
     ProductExchangeTable,
     TechnosphereExchangeTable,
 )
-from ...ui.tables.delegates import StringDelegate, FloatDelegate
 from ...ui.widgets import ActivityDataGrid, DetailsGroupBox, SignalledPlainTextEdit
 from ..panels.panel import ABTab
 
@@ -390,243 +386,6 @@ class ActivityTab(QtWidgets.QWidget):
                         del old_properties[property]
                 old_properties |= updated_properties
             self.activity.save()
-
-
-class PropertyModel(QtCore.QAbstractTableModel):
-    """
-    Model for the property editor table
-    The data is (str, float).
-    """
-
-    @dataclass
-    class PropertyData:
-        """Dataclass to represent one row with correct typehints"""
-        key: str = ""
-        value: float = 0
-        to_be_deleted: bool = False
-
-        def __getitem__(self, index: int) -> Union[str, float]:
-            """x[k] operator for easier usage"""
-            if index == 0:
-                return self.key
-            elif index == 1:
-                return self.value
-            raise IndexError
-
-        def __setitem__(self, index: int, value: Union[str, float]):
-            """x[k]=v operator for easier usage"""
-            if index == 0:
-                if not isinstance(value, str):
-                    raise TypeError
-                self.key = value
-            elif index == 1:
-                if not isinstance(value, float):
-                    raise TypeError
-                self.value = value
-            else:
-                raise IndexError
-
-
-    def __init__(self, read_only: bool):
-        super().__init__()
-        self._data: list[PropertyModel.PropertyData] = []
-        self._read_only = read_only
-
-    def populate(self, data: Optional[dict[str, float]]) -> None:
-        self._data.clear()
-        log.info(f"Input data for property table: {data}")
-        if data is not None:
-            self._original_data = copy(data)
-            for key in data:
-                self._data.append(PropertyModel.PropertyData(key, data[key]))
-        else:
-            self._original_data = dict()
-        self._data.append(PropertyModel.PropertyData())
-        self.layoutChanged.emit()
-
-    def data(self, index: QtCore.QModelIndex,
-             role: int = QtCore.Qt.ItemDataRole.DisplayRole) -> Any:
-        if index.isValid() and index.column() < 2:
-            if role == QtCore.Qt.ItemDataRole.EditRole:
-                return self._data[index.row()][index.column()]
-            elif (role == QtCore.Qt.ItemDataRole.DisplayRole
-                    or role == QtCore.Qt.ItemDataRole.ToolTipRole):
-                # Do not show values for properties which have no name
-                # to hint that these will not be saved
-                if (index.column() == 1 and self._data[index.row()][0] == ""):
-                    return nan
-                return self._data[index.row()][index.column()]
-            elif role == QtCore.Qt.ItemDataRole.FontRole:
-                font = QtGui.QFont()
-                if self._data[index.row()].to_be_deleted:
-                    font.setStrikeOut(True)
-                return font
-            elif role == QtCore.Qt.ItemDataRole.ForegroundRole:
-                current_key = self._data[index.row()].key
-                if self._data[index.row()].to_be_deleted:
-                    return style_item.brushes.get("deleted")
-                elif current_key in self._duplicate_keys():
-                    return style_item.brushes.get("duplicate")
-                elif self._original_data.get(current_key) == None:
-                    return style_item.brushes.get("new")
-                elif self._original_data.get(current_key) != self._data[index.row()].value:
-                    return style_item.brushes.get("modified")
-                return None
-        if (index.isValid() and index.column() == 2 
-                and (role == QtCore.Qt.ItemDataRole.DisplayRole 
-                     or role == QtCore.Qt.ItemDataRole.EditRole)):
-            return ""
-        return None
-
-    def setData(self, index: QtCore.QModelIndex, value: Any,
-             role: int = QtCore.Qt.ItemDataRole.DisplayRole) -> bool:
-        if (index.isValid() and role == QtCore.Qt.ItemDataRole.EditRole 
-                and index.column() < 2):
-            if self._data[index.row()][index.column()] != value:
-                self._data[index.row()][index.column()] = value
-                self.dataChanged.emit(index, index, [])
-            # Make sure there is an empty row at the end
-            if self._data[-1][0] != "":
-                self.beginInsertRows(QtCore.QModelIndex(), len(self._data), len(self._data))
-                self._data.append(PropertyModel.PropertyData())
-                self.endInsertRows()
-
-            return True
-        return False
-
-    def flags(self, index: QtCore.QModelIndex) -> QtCore.Qt.ItemFlags:
-        if index.isValid() and index.column() < 2:
-            result = (QtCore.Qt.ItemFlag.ItemIsSelectable
-                    | QtCore.Qt.ItemFlag.ItemIsEnabled) 
-            if not self._read_only:
-                result |= QtCore.Qt.ItemFlag.ItemIsEditable
-            return result
-        if index.isValid() and index.column() == 2:
-            return (QtCore.Qt.ItemFlag.ItemIsEnabled | QtCore.Qt.ItemFlag.ItemIsEditable) 
-        return None
-    
-    def rowCount(self, parent: QtCore.QModelIndex = QtCore.QModelIndex()) -> int:
-        return len(self._data)
-
-    def columnCount(self, parent: QtCore.QModelIndex = QtCore.QModelIndex()) -> int:
-        return 3
-
-    def headerData(self, section:int, orientation:QtCore.Qt.Orientation,
-                   role: int=QtCore.Qt.ItemDataRole.DisplayRole) -> Any:
-        
-        if (section < 3 and 
-                orientation == QtCore.Qt.Orientation.Horizontal
-                and role == QtCore.Qt.ItemDataRole.DisplayRole):
-            return ("Name", "Value", "")[section]
-        return None
-
-    def get_data_table(self) -> dict[str, float]:
-        result = { item.key:item.value for item in self._data if item.key != ""}
-        return result
-
-    def _duplicate_keys(self) -> list[str]:
-        duplicates:list[str] = []
-        key_set: set[str] = set()
-        for item in self._data:
-            if not item.to_be_deleted:            
-                if item.key in key_set:
-                    duplicates.append(item.key)
-                else:
-                    key_set.add(item.key)
-        return duplicates
-
-    def has_duplicate_key(self) -> bool:
-        return len(self._duplicate_keys()) > 0
-    
-    def handle_delete_request(self, index: QtCore.QModelIndex):
-        if index.isValid():
-            self._data[index.row()].to_be_deleted = not self._data[index.row()].to_be_deleted
-            start_index = self.createIndex(index.row(), 0)
-            end_index = self.createIndex(index.row(), 1)
-            self.dataChanged.emit(start_index, end_index, [])
-
-
-
-class DeleteButtonDelegate(QtWidgets.QStyledItemDelegate):
-    """For deleting rows"""
-
-    delete_request = QtCore.Signal(QtCore.QModelIndex)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-    def createEditor(self, 
-                     parent: QtWidgets.QWidget, 
-                     option: QtWidgets.QStyleOptionViewItem, 
-                     index: QtCore.QModelIndex):
-        button = QtWidgets.QPushButton()
-        button.setFocusPolicy(QtCore.Qt.FocusPolicy.NoFocus)
-        button.setFixedSize(20, 20)
-        button.setAutoDefault(False)
-        button.clicked.connect(lambda: self.delete_request.emit(index))
-        style = application.style()
-        icon = style.standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DockWidgetCloseButton)
-        button.setIcon(icon)
-        button.setStyleSheet("background-color: rgb(200, 150, 150); border-radius: 3")
-        editor = QtWidgets.QWidget(parent)
-        editor.setFocusPolicy(QtCore.Qt.FocusPolicy.NoFocus)
-        layout = QtWidgets.QHBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(button)
-        editor.setLayout(layout)
-        return editor
-
-
-
-class PropertyTable(QtWidgets.QTableView):
-    """Table view for editing properties"""
-    def __init__(self, model: PropertyModel, parent=None):
-        super().__init__(parent)
-        self.setVerticalScrollMode(QtWidgets.QTableView.ScrollPerPixel)
-        self.setHorizontalScrollMode(QtWidgets.QTableView.ScrollPerPixel)
-        self.setSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Expanding)
-        self.setWordWrap(True)
-        self.setAlternatingRowColors(True)
-        self.setSortingEnabled(False)
-        self.horizontalHeader().setHighlightSections(False)
-        self.horizontalHeader().setDefaultAlignment(QtCore.Qt.AlignLeft)
-        self.verticalHeader().setVisible(False)
-        self.setSelectionMode(QtWidgets.QAbstractItemView.SelectionMode.SingleSelection)
-        self.setEditTriggers(
-            QtWidgets.QAbstractItemView.EditTrigger.DoubleClicked |
-            QtWidgets.QAbstractItemView.EditTrigger.SelectedClicked |
-            QtWidgets.QAbstractItemView.EditTrigger.EditKeyPressed |
-            QtWidgets.QAbstractItemView.EditTrigger.AnyKeyPressed
-        )
-
-        self.setItemDelegateForColumn(0, StringDelegate(self))
-        # Use FloatDelegate, so that int values do not trigger an int validation
-        self.setItemDelegateForColumn(1, FloatDelegate(self))
-        delete_delegate = DeleteButtonDelegate(self)
-        self.setItemDelegateForColumn(2, delete_delegate)
-        # self.setItemDelegateForColumn(2, CheckboxDelegate(self))
-
-        self._model = model
-        self.setModel(self._model)
-        delete_delegate.delete_request.connect(self._model.handle_delete_request)
-        self._model.rowsInserted.connect(self._handle_rows_inserted)
-
-        self.horizontalHeader().setSectionResizeMode(1, QtWidgets.QHeaderView.ResizeMode.Stretch)
-        self.horizontalHeader().setSectionResizeMode(2, QtWidgets.QHeaderView.ResizeMode.Fixed)
-        self.horizontalHeader().resizeSection(2, 40)
-
-
-    def populate(self, data: Optional[dict[str, float]]) -> None:
-        self._model.populate(data)
-        for i in range(self._model.rowCount()):
-            index = self._model.createIndex(i, 2)
-            self.openPersistentEditor(index)
-
-    def _handle_rows_inserted(self, parent: QtCore.QModelIndex, first: int, last: int):
-        # first , last are inclusive
-        for i in range(first, last + 1):
-            index = self._model.createIndex(i, 2)
-            self.openPersistentEditor(index)
